Add an option to control the file path in the saved bookmark file
g:bookmark_relative_path by default is 0, and the full file path is stored.
When set to 1, only the relative to the current working directory path is
stored.

The reason to do that is the folder and its upper parent folders can be
changed or moved to other location, and the bookmark file will still be
valid.